Test envfile functions
